// src/dashboard/types/RecipeData.ts
interface IRecipeData {
   _id: string;
   source: string;
   author: string;
   title: string;
   description: string;
   imageUrl: string;
   url: string;
   tags: string[];
   comments: string[];
   userUuid: string;
   addTag: (tag: string) => void;
   __v: number;
}

export default IRecipeData;